requirements.txt
sqlalchemy-2.0.7
mysql-connector-python-8.0.32
pymysql-1.0.2
waitress-2.1.2
flask-2.2.3
flask_bcrypt-1.0.1
requests-2.28.2
numpy-1.24.2
flask_wtf-1.1.1
email_validator-1.3.1
pandas-1.5.3
cryptography-41.0.2
